updated api validations

+++ api/v1/db/client.js
@@ -68,19 +68,22 @@ const deleteClient = async (clientId) => {
 			rowMode: "array",
 		};
 		let resp = await query(_sql);
-
-		let claimsArr = resp?.rows && resp?.rows?.map((item) => item[0]);
-		_sqltext = `delete from insurance_claim where insurance_policy_id in (${claimsArr})`;
-		await query(_sqltext);
-
+		console.log(resp, "resp");
+		let claimsArr = resp?.rows?.length && resp?.rows?.map((item) => item[0]);
+		if (claimsArr) {
+			_sqltext = `delete from insurance_claim where insurance_policy_id in (${claimsArr})`;
+			resp = await query(_sqltext);
+			console.log(resp, "resp 2", claimsArr);
+		}
 		_sqltext = `delete from insurance_policy where client_id=$1`;
-		await query(_sqltext, clientId);
-
+		resp = await query(_sqltext, clientId);
+		console.log(resp, "resp 3");
 		_sqltext = `delete from client where id = $1`;
-		await query(_sqltext, clientId);
-
+		resp = await query(_sqltext, clientId);
+		console.log(resp, "resp 4");
 		return "deleted";
 	} catch (error) {
+		console.log(error, "eror");
 		throw error;
 	}
 };
@@ -89,6 +92,13 @@ const createUser = async (reqBody) => {
 	const { query } = db.createConnection();
 	const { email, password } = reqBody;
 	try {
+		const sql = {
+			name: "delete-user",
+			text: `delete from users where email=$1`,
+			values: [email],
+		};
+		await query(sql);
+
 		const _sql = {
 			name: "create-user",
 			text: `insert into users(email,password) values ($1,$2) returning *`,
@@ -115,6 +125,37 @@ const getUserByEmail = async (email, password) => {
 		throw error;
 	}
 };
+
+const getUserStatus = async (id) => {
+	const { query } = db.createConnection();
+	try {
+		const _sql = {
+			name: "get-user-status",
+			text: `select login_status from users where id=$1`,
+			values: [id],
+		};
+		const data = await query(_sql);
+		return data.rows[0];
+	} catch (error) {
+		throw error;
+	}
+};
+
+const updateLoginStatus = async (id) => {
+	const { query } = db.createConnection();
+	try {
+		const _sql = {
+			name: "get-user-status",
+			text: `update users set login_status=$1 where id=$2`,
+			values: [true, id],
+		};
+		await query(_sql);
+		return "done";
+	} catch (error) {
+		throw error;
+	}
+};
+
 const client = {
 	getAllClients,
 	getClientById,
@@ -123,6 +164,8 @@ const client = {
 	deleteClient,
 	createUser,
 	getUserByEmail,
+	getUserStatus,
+	updateLoginStatus,
 };
 
 module.exports = client;
